move ParseError to own file
